# examples/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Spectra Test Backend API",
    "description": "Comprehensive API for testing the Spectra testing framework",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
...
    "/todos": {
      "get": {
        "summary": "Get all todos",
        "description": "Retrieve all todos",
        "operationId": "getTodos",
        "responses": {
          "200": {
            "description": "List of todos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new todo",
        "description": "Add a new todo",
        "operationId": "createTodo",
        "requestBody": {
          "description": "Todo to add",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Todo created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      },
      "patch": {
        "summary": "Update all todos completion status",
        "description": "Toggle completion status for all todos",
        "operationId": "updateAllTodos",
        "requestBody": {
          "description": "Completion status",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["completed"],
                "properties": {
                  "completed": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todos updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodosUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/todos/{id}": {
      "get": {
        "summary": "Get todo by ID",
        "description": "Retrieve a specific todo by its ID",
        "operationId": "getTodoById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Todo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Todo details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "404": {
            "description": "Todo not found"
          }
        }
      },
      "put": {
        "summary": "Update a todo",
        "description": "Update an existing todo",
        "operationId": "updateTodo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Todo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Todo data to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Todo not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a todo",
        "description": "Remove a todo by its ID",
        "operationId": "deleteTodo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Todo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Todo deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Todo not found"
          }
        }
      }
    },
...
    "/weather/delayed": {
      "get": {
        "summary": "Get weather data with delay",
        "description": "Retrieve weather data with a simulated delay",
        "operationId": "getDelayedWeather",
        "parameters": [
          {
            "name": "delay",
            "in": "query",
            "description": "Delay in seconds (1-10)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10,
              "default": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delayed weather data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/WeatherData"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/weather/error": {
      "get": {
        "summary": "Simulated error endpoint",
        "description": "Endpoint that returns the specified error code",
        "operationId": "getWeatherError",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "HTTP error code (400-599)",
            "schema": {
              "type": "integer",
              "minimum": 400,
              "maximum": 599,
              "default": 500
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Todo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "completed": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TodoCreateRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Buy groceries"
          },
          "completed": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "TodoUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "completed": {
            "type": "boolean"
          }
        }
      },
      "TodosUpdateResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "todos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Todo"
            }
          }
        }
      },
      "WeatherData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "temperature": {
            "type": "number",
            "format": "float"
          },
          "conditions": {
            "type": "string"
          },
          "humidity": {
            "type": "number"
          },
          "windSpeed": {
            "type": "number",
            "format": "float"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer"
          }
        }
      }
    }
  }
}